Timelimit is u32 and not u64

+++ src/builder.rs
@@ -633,12 +633,12 @@ pub struct Builder {
     bundle_type: BundleType,
     anchor: Anchor,
     // When timelimit is set, the Builder will build an ActionGroup (burn must be empty)
-    timelimit: Option<u64>,
+    timelimit: Option<u32>,
 }
 
 impl Builder {
     /// Constructs a new empty builder for an Orchard bundle.
-    pub fn new(bundle_type: BundleType, anchor: Anchor, timelimit: Option<u64>) -> Self {
+    pub fn new(bundle_type: BundleType, anchor: Anchor, timelimit: Option<u32>) -> Self {
         Builder {
             spends: vec![],
             outputs: vec![],
@@ -1006,7 +1006,7 @@ pub fn bundle<V: TryFrom<i64>, FL: OrchardFlavor>(
 fn build_bundle<B, R: RngCore>(
     mut rng: R,
     anchor: Anchor,
-    timelimit: Option<u64>,
+    timelimit: Option<u32>,
     bundle_type: BundleType,
     spends: Vec<SpendInfo>,
     outputs: Vec<OutputInfo>,
